PIM-4940: Move UserType from Oro to Pim
