+++ features/Context/WebUser.php
@@ -2544,7 +2544,7 @@ class WebUser extends RawMinkContext
      */
     public function iShouldSeeLocaleOption($not, $locale)
     {
-        $selectNames = ['localization[oro_locale___language][value]', 'oro_user_user_form[uiLocale]'];
+        $selectNames = ['localization[oro_locale___language][value]', 'pim_user_user_form[uiLocale]'];
         $field = null;
         foreach ($selectNames as $selectName) {
             $field = (null !== $field) ? $field : $this->getCurrentPage()->findField($selectName);
