+++ src/Pim/Bundle/UserBundle/Resources/config/view_elements/user.yml
@@ -12,7 +12,7 @@ services:
             - 'pim_user.user.tab.general'
             - %pim_user.view_element.user.tab.general.template%
         tags:
-            - { name: pim_enrich.view_element, type: oro_user_user_form.form_tab, position: 90 }
+            - { name: pim_enrich.view_element, type: pim_user_user_form.form_tab, position: 90 }
 
     pim_user.view_element.user.tab.additional:
         parent: pim_enrich.view_element.base
@@ -20,7 +20,7 @@ services:
             - 'pim_user.user.tab.additional'
             - %pim_user.view_element.user.tab.additional.template%
         tags:
-            - { name: pim_enrich.view_element, type: oro_user_user_form.form_tab, position: 100 }
+            - { name: pim_enrich.view_element, type: pim_user_user_form.form_tab, position: 100 }
 
     pim_user.view_element.user.tab.group_and_role:
         parent: pim_enrich.view_element.base
@@ -30,7 +30,7 @@ services:
         calls:
             - [ addVisibilityChecker, ['@pim_user.view_element.visibility_checker.groups_and_roles'] ]
         tags:
-            - { name: pim_enrich.view_element, type: oro_user_user_form.form_tab, position: 110 }
+            - { name: pim_enrich.view_element, type: pim_user_user_form.form_tab, position: 110 }
 
     pim_user.view_element.user.tab.password:
         parent: pim_enrich.view_element.base
@@ -40,7 +40,7 @@ services:
         calls:
             - [ addVisibilityChecker, ['@pim_enrich.view_element.visibility_checker.non_empty_property', {property: '[form].children[change_password]'}] ]
         tags:
-            - { name: pim_enrich.view_element, type: oro_user_user_form.form_tab, position: 120 }
+            - { name: pim_enrich.view_element, type: pim_user_user_form.form_tab, position: 120 }
 
     pim_user.view_element.user.tab.interfaces:
         parent: pim_enrich.view_element.base
@@ -48,4 +48,4 @@ services:
             - 'pim_user.user.tab.interfaces'
             - %pim_user.view_element.user.tab.interfaces.template%
         tags:
-            - { name: pim_enrich.view_element, type: oro_user_user_form.form_tab, position: 130 }
+            - { name: pim_enrich.view_element, type: pim_user_user_form.form_tab, position: 130 }
